requirements.txt
groq>=0.9.0
pinecone>=5.0.0
sentence-transformers>=2.7.0
pypdf>=4.0.0
python-docx>=1.1.0
streamlit>=1.35.0
numpy>=1.26.0
python-dotenv>=1.0.0